+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Behaviour changes:
 Improvements:
  * minimum required Go version is 1.22 (#122, #126)
  * log Commit signature verification error (#134)
+ * add Commit message verification callback (#134)
 
 Bugs fixed:
  * context-bound PreBlock and PreHeader are not reset properly (#127)